8 commit

+++ application/controllers/Publication_emotions.php
@@ -9,7 +9,9 @@ class Publication_emotions extends CI_Controller {
         $this->load->model('users_model');
     }
 
-    public function get_publication_emotions() {
+    public function Index() {
+        $this->load->view('session_user');
+
         $publication_id = $this->input->post('publication_id');
         $publication_emotions = $this->publications_model->getPublicationEmotionsByPublicationId($publication_id);
         $html = '';
@@ -18,7 +20,7 @@ class Publication_emotions extends CI_Controller {
             $html .= "<div class='col-xs-6 col-sm-4 col-lg-3 emotion_user'>
                         <a href='" . base_url() . "one_user/$publication_emotion->email'>
                             <div class='emotion_user_image'>
-                                <img src='" . base_url() . "uploads/images/user_images/$publication_emotion->main_image' class='emotion_avatar' style='width: 100px;'>
+                                <img src='" . base_url() . "uploads/images/user_images/$publication_emotion->main_image' class='action_avatar' style='width: 100px;'>
                             </div>
                             <div class='emotion_user_name'>
                                 $publication_emotion->nickname $publication_emotion->surname
@@ -27,14 +29,16 @@ class Publication_emotions extends CI_Controller {
                     </div>";
         }
         $html .= "</div>";
-        $data = array(
-            'publication_emotions' => $html,
+        $get_emotions_json = array(
+            'one_publication_emotions' => $html,
             'csrf_hash' => $this->security->get_csrf_hash()
         );
-        echo json_encode($data);
+        echo json_encode($get_emotions_json);
     }
 
     public function insert_publication_emotion() {
+        $this->load->view('session_user');
+        $session_user_id = $_SESSION['user_id'];
         $emotion_date = date('d.m.Y');
         $emotion_time = date('H:i:s');
         $published_user_id = $this->input->post('published_user_id');
@@ -42,13 +46,13 @@ class Publication_emotions extends CI_Controller {
         $publication_id = $this->input->post('publication_id');
         $emotion_num_rows = $this->publications_model->getPublicationEmotionNumRowsByPublicationIdAndEmotionedUserId($publication_id, $emotioned_user_id);
 
-        if ($emotion_num_rows == 0) {
+        if ($emotion_num_rows == 0 && $emotioned_user_id == $session_user_id) {
             $data_publication_emotions = array(
                 'emotion_date' => $emotion_date,
                 'emotion_time' => $emotion_time,
                 'published_user_id' => $published_user_id,
                 'emotioned_user_id' => $emotioned_user_id,
-                'publication_id' => $publication_id,
+                'publication_id' => $publication_id
             );
             $this->publications_model->insertPublicationEmotion($data_publication_emotions);
 
@@ -63,20 +67,20 @@ class Publication_emotions extends CI_Controller {
                 'notification_date' => $emotion_date,
                 'notification_time' => $emotion_time,
                 'notification_viewed' => 'Не просмотрено',
+                'link_id' => $publication_id,
+                'link_table' => 'publications',
                 'user_id' => $published_user_id
             );
             $this->users_model->insertUserNotification($data_user_notifications);
             $insert_json = array(
                 'emotion_num_rows' => $emotion_num_rows,
                 'total_emotions' => $total_emotions,
-                'csrf_name' => $this->security->get_csrf_token_name(),
                 'csrf_hash' => $this->security->get_csrf_hash()
             );
         } else {
             $insert_json = array(
                 'emotion_num_rows' => $emotion_num_rows,
-                'emotion_error' => "Вы уже ставили эмоцию на данную публикацию",
-                'csrf_name' => $this->security->get_csrf_token_name(),
+                'emotion_error' => "Вы уже ставили эмоцию на данную публикацию или что-то пошло не так!",
                 'csrf_hash' => $this->security->get_csrf_hash()
             );
         }
@@ -84,10 +88,12 @@ class Publication_emotions extends CI_Controller {
     }
 
     public function delete_publication_emotion() {
+        $this->load->view('session_user');
+        $session_user_id = $_SESSION['user_id'];
         $publication_id = $this->input->post('publication_id');
         $emotioned_user_id = $this->input->post('emotioned_user_id');
         $emotion_num_rows = $this->publications_model->getPublicationEmotionNumRowsByPublicationIdAndEmotionedUserId($publication_id, $emotioned_user_id);
-        if ($emotion_num_rows > 0) {
+        if ($emotion_num_rows > 0 && $emotioned_user_id == $session_user_id) {
             $this->publications_model->deletePublicationEmotionByPublicationIdAndEmotionedUserId($publication_id, $emotioned_user_id);
             $total_emotions = $this->publications_model->getTotalByPublicationIdAndPublicationTable($publication_id, 'publication_emotions');
             $delete_json = array(
@@ -98,7 +104,7 @@ class Publication_emotions extends CI_Controller {
         } else {
             $delete_json = array(
                 'emotion_num_rows' => $emotion_num_rows,
-                'emotion_error' => "Вы ещё не ставили эмоцию на данную публикацию",
+                'emotion_error' => "Вы ещё не ставили эмоцию на данную публикацию или что-то пошло не так",
                 'csrf_hash' => $this->security->get_csrf_hash()
             );
         }
